# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Brand Envoy Africa API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactInputTier = typeof ContactInputTier[keyof typeof ContactInputTier];


export const ContactInputTier = {
  sme: 'sme',
  enterprise: 'enterprise',
  campaign: 'campaign',
} as const;

/**
 * @nullable
 */
export type ContactInputOfficeLevel = typeof ContactInputOfficeLevel[keyof typeof ContactInputOfficeLevel] | null;


export const ContactInputOfficeLevel = {
  local_government: 'local_government',
  state_assembly: 'state_assembly',
  senatorial: 'senatorial',
  gubernatorial: 'gubernatorial',
  presidential: 'presidential',
} as const;

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  phone?: string | null;
  tier: ContactInputTier;
  /** @minLength 1 */
  message: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  market?: string | null;
  /** @nullable */
  officeLevel?: ContactInputOfficeLevel;
  /** @nullable */
  services?: string[] | null;
}

export type ContactSubmissionTier = typeof ContactSubmissionTier[keyof typeof ContactSubmissionTier];


export const ContactSubmissionTier = {
  sme: 'sme',
  enterprise: 'enterprise',
  campaign: 'campaign',
} as const;

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  phone?: string | null;
  tier: ContactSubmissionTier;
  message: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  market?: string | null;
  /** @nullable */
  officeLevel?: string | null;
  /** @nullable */
  services?: string[] | null;
  createdAt: string;
}

export interface AffiliateInput {
  /** @minLength 1 */
  productName: string;
  affiliateUrl: string;
  /** @minLength 1 */
  productInfo: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  published?: boolean;
}

export interface AffiliateUpdate {
  productName?: string;
  affiliateUrl?: string;
  productInfo?: string;
  /** @nullable */
  generatedCopy?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  published?: boolean;
}

export interface AffiliateProduct {
  id: number;
  productName: string;
  affiliateUrl: string;
  productInfo: string;
  /** @nullable */
  generatedCopy?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PublishToggleInput {
  published: boolean;
}

export type AffiliateStatsCategoriesItem = {
  category: string;
  count: number;
};

export interface AffiliateStats {
  total: number;
  published: number;
  unpublished: number;
  categories: AffiliateStatsCategoriesItem[];
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  date: string;
  categories?: string[];
  /** @nullable */
  featuredImageUrl?: string | null;
  /** @nullable */
  author?: string | null;
}

export interface BlogCategory {
  id: number;
  name: string;
  count: number;
}

export interface BlogPostList {
  posts: BlogPost[];
  total: number;
  page: number;
  perPage: number;
}

export type ListContactsParams = {
tier?: ListContactsTier;
limit?: number;
offset?: number;
};

export type ListContactsTier = typeof ListContactsTier[keyof typeof ListContactsTier];


export const ListContactsTier = {
  sme: 'sme',
  enterprise: 'enterprise',
  campaign: 'campaign',
} as const;

export type ListAffiliatesParams = {
publishedOnly?: boolean;
category?: string;
limit?: number;
offset?: number;
};

export type ListBlogPostsParams = {
page?: number;
perPage?: number;
category?: string;
search?: string;
};