# lib/api-zod/src/generated/types/affiliateStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Brand Envoy Africa API
 * OpenAPI spec version: 0.1.0
 */
import type { AffiliateStatsCategoriesItem } from './affiliateStatsCategoriesItem';

export interface AffiliateStats {
  total: number;
  published: number;
  unpublished: number;
  categories: AffiliateStatsCategoriesItem[];
}