# lib/api-zod/src/generated/types/affiliateUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Brand Envoy Africa API
 * OpenAPI spec version: 0.1.0
 */

export interface AffiliateUpdate {
  productName?: string;
  affiliateUrl?: string;
  productInfo?: string;
  /** @nullable */
  generatedCopy?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  published?: boolean;
}