# lib/api-zod/src/generated/types/blogCategory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Brand Envoy Africa API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogCategory {
  id: number;
  name: string;
  count: number;
}