# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Brand Envoy Africa API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  date: Date;
  categories?: string[];
  /** @nullable */
  featuredImageUrl?: string | null;
  /** @nullable */
  author?: string | null;
}